+++ src/IRegex.ts
@@ -1,19 +1,15 @@
-export interface RegexSettings {
+export interface RXData {
+    [key: string]: string[] | string
+}
+
+export interface RXPlaceholder {
+    [key: string]: string[]
+}
+
+export interface RXSettings {
     template: string
     flags: string
+    symbol?: string
+    templateList?: string[]
 }
 
-export interface RegexData {
-    settings: RegexSettings,
-    [key: string]: any
-}
-
-export interface RegexListSettings {
-    template: Array<string>
-    flags: string
-}
-
-export interface RegexListData {
-    settings: RegexListSettings
-    [key: string]: any
-}
