+++ src/builderbase.ts
@@ -1,23 +1,4 @@
-import { AugmentedExp } from './augmentedExp';
-
-interface Buildable {
-    build() : RegExp;
-}
-
-export interface RXData {
-    [key: string]: string[] | string
-}
-
-export interface RXPlaceholder {
-    [key: string]: string[]
-}
-
-export interface RXSettings {
-    template: string
-    flags: string
-    symbol?: string
-    templateList?: string[]
-}
+import { RXData, RXSettings, RXPlaceholder } from './IRegex';
 
 export class RegexBuilderBase {
     protected _regexData: RXData;
@@ -50,36 +31,6 @@ export class RegexBuilderBase {
             if (!thisRef._placeholderData[p1]) throw new Error(`found undefined placeholder ${match} in regex data`);
             return thisRef._buildGroup(thisRef._placeholderData[p1]);
         });
-    }
-
-}
-
-export class TemplateBuilder extends RegexBuilderBase {
-
-    public build() : AugmentedExp {
-        return this._buildRegex(this._buildTemplate(this._settings.template));
-    }
-
-    private _buildRegex(template: string) : AugmentedExp {
-        return new AugmentedExp(template, this._settings.flags, this._settings.template);
-    }
-
-}
-
-export class TemplateListBuilder extends RegexBuilderBase {
-
-    public build() : Array<RegExp> {
-        if (!this._settings.templateList) throw new Error('templateList must be defined');
-        const regexList: Array<RegExp> = [];
-        for (let template of this._settings.templateList) {
-            regexList.push(this._buildRegex(this._buildTemplate(template)));
-        }
-        
-        return regexList;
-    }
-
-    private _buildRegex(template: string) : RegExp {
-        return new RegExp(template, this._settings.flags);
     }
 
 }
