+++ src/index.ts
@@ -1,6 +1,6 @@
-export { RegexBuilder } from './regexbuilder'
-export { TemplateMapper } from './templatemapper'
-export { RegexJSONBuilder } from './rjsonbuilder'
-export { RegexListBuilder } from './listbuilder'
-export { TemplateBuilder, TemplateListBuilder, RXData, RXSettings, RXPlaceholder } from './buildernew'
+export { RegexBuilder } from './regexbuilder';
+export { RXData, RXSettings, RXPlaceholder } from './IRegex';
+export { TemplateBuilder, TemplateListBuilder} from './templatebuilder';
+export { TemplateMapper } from './templatemapper';
+export { AugmentedExp } from './augmentedExp';
 export { Regexpress } from './regexpress';
